Fix bot not finding config file
The bot previously would previously look for the config file
using a path relative to the current working directory. Now
it uses a path relative to its own directory. This means that
the bot can now be executed from outside its directory.

In the future, when loading files, developers should ensure
the files are loaded using a path relative to util.basePath